requirements.txt
omegaconf>=2.3.0
hydra-core>=1.3.2
hydra-callbacks>=0.6.1
joblib>=1.4.2
scikit-learn>=1.5.1
pydicom==2.4.4
nibabel==5.2.1
pylidc==0.2.3
imops==0.9.0
medimm==0.0.0